backend/requirements.txt
Flask
Flask-Cors
mysql-connector-python
requests
beautifulsoup4
PyPDF2
pandas
